apps/api/pyproject.toml
[project]
name = "llm-excel-api"
version = "0.1.0"
description = "LLM辅助生成Excel数据处理规则的系统 - API服务"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "sse-starlette>=2.0.0",
    # 数据库
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # 认证
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.0.0",  # Pydantic EmailStr 需要
]
[project.scripts]
dev = "uvicorn llm_excel_api.main:app --reload --host localhost --port 8000"

